Move VMOffset (and a few friends) to wasmer_types, and adapted compiler-singlepass to use it
