+++ lib/compiler-singlepass/src/codegen.rs
@@ -12,13 +12,10 @@ use wasmer_compiler::{
 };
 use wasmer_types::{
     entity::{EntityRef, PrimaryMap, SecondaryMap},
-    FunctionType,
+    FunctionIndex, FunctionType, GlobalIndex, LocalFunctionIndex, LocalMemoryIndex, MemoryIndex,
+    ModuleInfo, SignatureIndex, TableIndex, TrapCode, Type, VMBuiltinFunctionIndex, VMOffsets,
 };
-use wasmer_types::{
-    FunctionIndex, GlobalIndex, LocalFunctionIndex, LocalMemoryIndex, MemoryIndex, ModuleInfo,
-    SignatureIndex, TableIndex, TrapCode, Type,
-};
-use wasmer_vm::{MemoryStyle, TableStyle, VMBuiltinFunctionIndex, VMOffsets};
+use wasmer_vm::{MemoryStyle, TableStyle};
 
 /// The singlepass per-function code generator.
 pub struct FuncGen<'a, M: Machine> {
