+++ lib/compiler-singlepass/src/machine_x64.rs
@@ -13,8 +13,7 @@ use wasmer_compiler::{
     InstructionAddressMap, Relocation, RelocationKind, RelocationTarget, SectionBody, SourceLoc,
     TrapInformation,
 };
-use wasmer_types::{FunctionIndex, FunctionType, TrapCode, Type};
-use wasmer_vm::VMOffsets;
+use wasmer_types::{FunctionIndex, FunctionType, TrapCode, Type, VMOffsets};
 
 type Assembler = VecAssembler<X64Relocation>;
