+++ lib/types/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 [dependencies]
 serde = { version = "1.0", features = ["derive", "rc"], optional = true, default-features = false }
 thiserror = "1.0"
+more-asserts = "0.2"
 indexmap = { version = "1.6", features = ["serde-1"] }
 rkyv = { version = "0.7.20", optional = true }
 loupe = { version = "0.1", features = ["enable-indexmap"] }
