+++ lib/vm/Cargo.toml
@@ -12,7 +12,6 @@ edition = "2018"
 
 [dependencies]
 wasmer-types = { path = "../types", version = "=2.2.1" }
-region = "3.0"
 libc = { version = "^0.2", default-features = false }
 memoffset = "0.6"
 indexmap = { version = "1.6", features = ["serde-1"] }
@@ -24,9 +23,12 @@ serde = { version = "1.0", features = ["derive", "rc"] }
 rkyv = { version = "0.7.20", optional = true }
 loupe = { version = "0.1", features = ["enable-indexmap"] }
 enum-iterator = "0.7.0"
-corosensei = { version = "0.1.2" }
 scopeguard = "1.1.0"
 lazy_static = "1.4.0"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+region = { version = "3.0" }
+corosensei = { version = "0.1.2" }
 
 [target.'cfg(target_vendor = "apple")'.dependencies]
 mach = "0.3.2"
